refactor: geocode/file 모듈 분리 (#33)
