+++ src/main/java/goorm/_44/controller/geocode/GeocodingController.java
@@ -1,15 +1,16 @@
-package goorm._44.controller.map;
+package goorm._44.controller.geocode;
 
 import goorm._44.config.api.ApiResult;
-import goorm._44.service.map.GeocodingService;
+import goorm._44.dto.response.GeocodeResponse;
+import goorm._44.service.geocode.GeocodingService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import org.springframework.web.bind.annotation.*;
 import reactor.core.publisher.Mono;
 
 @RestController
-@RequestMapping("/geocode")
-@Tag(name = "Geocoding", description = "좌표 ↔ 주소 변환 API")
+@RequestMapping("/api/geocode")
+@Tag(name = "Geocode", description = "위경도 주소 관련 API")
 public class GeocodingController {
 
     private final GeocodingService geocodingService;
@@ -19,15 +20,12 @@ public class GeocodingController {
     }
 
     @GetMapping("/address")
-    @Operation(summary = "좌표로 주소 조회", description = "위도(lat), 경도(lng)로 도로명/지번 주소를 반환합니다.")
+    @Operation(summary = "위경도 변환", description = "위도(lat), 경도(lng)를 주소로 변환합니다.")
     public Mono<ApiResult<GeocodeResponse>> getAddress(
             @RequestParam double lat,
             @RequestParam double lng
     ) {
         return geocodingService.getAddressFromCoordinates(lat, lng)
-                .map(addr -> ApiResult.success(new GeocodeResponse(addr)));
+                .map(opt -> ApiResult.success(new GeocodeResponse(opt.orElse(null))));
     }
-
-    // 응답 바디 DTO
-    public record GeocodeResponse(String address) {}
 }
